Redpanda/py_run/requirements.txt
streamlit
requests
geopandas
kafka-python
fiona
shapely
rtree